requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///C:/ci/aiohttp_1637857271452/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
astor==0.8.1
astunparse==1.6.3
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
awsebcli==3.20.3
blinker==1.4
boto3==1.21.4
botocore==1.24.5
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
cement==2.8.2
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
colorama==0.4.3
cryptography @ file:///C:/ci/cryptography_1633520552480/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-cors-headers==3.11.0
django-ninja==0.17.0
dlib @ file:///D:/bld/dlib_1645480142555/work
fonttools==4.25.0
frozenlist @ file:///C:/ci/frozenlist_1637767289747/work
future==0.16.0
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth @ file:///tmp/build/80754af9/google-auth_1626320605116/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///C:/ci/grpcio_1637590968244/work
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1638543108096/work
jmespath==0.10.0
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///C:/ci/kiwisolver_1644962567532/work
Markdown @ file:///C:/ci/markdown_1614364121613/work
matplotlib @ file:///C:/ci/matplotlib-suite_1645455836921/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
multidict @ file:///C:/ci/multidict_1607362065515/work
munkres==1.1.4
numpy==1.22.2
oauthlib @ file:///opt/conda/conda-bld/oauthlib_1644332107998/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.5.62
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pathspec==0.9.0
Pillow==8.4.0
protobuf==3.19.4
psycopg2 @ file:///C:/ci/psycopg2_1612298803839/work
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.9.0
PyJWT @ file:///C:/ci/pyjwt_1619651841937/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pypiwin32==223
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pywin32==303
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
s3transfer==0.5.1
scipy==1.8.0
semantic-version==2.8.5
sip==4.19.13
six==1.14.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator @ file:///tmp/build/80754af9/tensorflow-estimator_1599136169057/work/whl_temp/tensorflow_estimator-2.3.0-py2.py3-none-any.whl
tensorflow-hub==0.12.0
termcolor==1.1.0
tornado @ file:///C:/ci/tornado_1606942392901/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wcwidth==0.1.9
Werkzeug @ file:///tmp/build/80754af9/werkzeug_1635505089296/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt @ file:///C:/ci/wrapt_1638434029846/work
yarl @ file:///C:/ci/yarl_1606940076464/work
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
